refactor view model to use case model
